requirements.txt
aiohttp
asyncio
async-timeout
nextcord.py
gTTS
IMDbPY
kaleido
numpy 
opencv-python 
plotly
pybase64
python-dotenv
python-osrsapi
requests
yfinance
youtube_dl
